Memento-S/pyproject.toml
[project]
name = "memento-s"
version = "0.1.0"
description = "Memento-S Agent CLI"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "anthropic",
    "openai",
    "litellm",
    "mcp",
    "fastmcp",
    "tiktoken",
    "httpx",
    "aiohttp",
    "requests",
    "aiofiles",
    "anyio",
    "pyyaml",
    "google_search_results",
    "crawl4ai",
    "camel-ai",
    "nltk",
    "prompt_toolkit",
    "tqdm",
    "pygame>=2.6.1",
]

[project.scripts]
memento = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "core"]
include = [
    "agent.py",
    "AGENTS.md",
    "skills/**",
]

[tool.hatch.build.targets.sdist]
include = [
    "cli/",
    "core/",
    "skills/",
    "agent.py",
    "AGENTS.md",
    "pyproject.toml",
    "README.md",
]
